Finish IO factor for now
The io_logical crate is still embedded in this
repository. When it's stabilized, and needed in
another project, we can publish it under its own
repository.
