+++ io-logical/Cargo.toml
@@ -11,3 +11,6 @@ thiserror       = "1.0"
 serial_test     = "0.5"
 glob            = "0.3"
 either          = "1.6"
+num             = "0.3"
+ndarray         = "0.13"
+hdf5            = "0.7"
